Extract dismissWarningDialogs into helpers/report-utils.ts
The 38-line helper was copied verbatim into both
cm-escape-rate-display.ts and cm-escape-rate-predictive-borderline.ts.
Two copies will eventually drift if the warning-dialog locators or the
modal classes change. Hoist the function into helpers/report-utils.ts
where it can be shared, and replace both test-local definitions with
an import.

The hoisted version also gains a longer docstring that explains why the
helper exists at all (warnings/acks stack across resolved reports and
block subsequent interactions) and why the loop bound of 10 is a
defensive cap rather than a precise count — context that wasn't in
either test-local copy.
