+++ e2e-tests/helpers/report-utils.ts
@@ -204,3 +204,57 @@ export async function withReportCountTracking<T>(
         timeoutMs,
     );
 }
+
+/**
+ * Dismiss any warning/ack dialogs that have accumulated on a user's page.
+ *
+ * After a CM vote that issues a warning the affected player sees a modal
+ * (`.AccountWarning` for formal, `.AccountWarningInfo` for informal); the
+ * reporter sees an acknowledgement modal (`.AccountWarningAck`). These
+ * stack across multiple resolved reports and block subsequent interactions
+ * (e.g. accepting the next challenge), so tests that resolve several
+ * reports in sequence need to drain them between iterations.
+ *
+ * The loop bound of 10 is a defensive cap: in practice there is at most
+ * one dialog per resolved report, but we keep iterating until `waitFor`
+ * times out so the helper is robust to whatever stack depth exists.
+ */
+export async function dismissWarningDialogs(page: Page): Promise<void> {
+    // Dismiss formal warnings (require checking "I understand" checkbox)
+    const formalWarning = page.locator("div.AccountWarning");
+    for (let i = 0; i < 10; i++) {
+        try {
+            await formalWarning.waitFor({ state: "visible", timeout: 3000 });
+            const checkbox = formalWarning.locator('input[type="checkbox"]');
+            await checkbox.check();
+            await formalWarning.locator("button.primary").click();
+            await expect(formalWarning).not.toBeVisible();
+        } catch {
+            break;
+        }
+    }
+
+    // Dismiss informal warnings
+    const infoOk = page.locator(".AccountWarningInfo button.primary");
+    for (let i = 0; i < 10; i++) {
+        try {
+            await infoOk.waitFor({ state: "visible", timeout: 3000 });
+            await infoOk.click();
+            await expect(infoOk).not.toBeVisible();
+        } catch {
+            break;
+        }
+    }
+
+    // Dismiss ack dialogs (reporter gets these)
+    const ackOk = page.locator("div.AccountWarningAck button.primary");
+    for (let i = 0; i < 10; i++) {
+        try {
+            await ackOk.waitFor({ state: "visible", timeout: 3000 });
+            await ackOk.click();
+            await expect(ackOk).not.toBeVisible();
+        } catch {
+            break;
+        }
+    }
+}
